Fix data loss in the peer stores, NVM transfer completion and UART drop counts
Storage
-------

bt_pds narrowed the record sequence to 16 bits when setting s_NextSeq at
mount, while Seq is uint32_t in the record header and both comparisons
that pick the newest record use the full width. Past 65535 records the
counter restarted low on the next mount and a record written afterwards
lost to an older one in FindLatest.

NvmDiskIO::SectWrite returned as soon as Nvm::Write answered. In
interrupt mode that is when the first chunk has been handed over, with
the rest still to be read out of the caller's buffer, and nvm.h requires
that buffer to stay valid until the completion event or Sync. FatFS and
littlefs both reuse their sector buffer immediately. Erase and
EraseSector wait as well, since nothing follows them to drain the medium.

Nvm::IntrfEvent folded TX_READY and TX_FIFO_EMPTY into completion.
TX_READY asks the handler for more data and TX_FIFO_EMPTY says a software
FIFO drained; neither means the transfer finished. The nRF I2C master
raises TX_READY on EVENTS_LASTTX in its DMA path, before the stop
condition, and Nvm supports I2C EEPROM as its bare medium case, so a
chunk was returned to the driver while the bus was still on it. Every
interface Nvm can be given was checked: the nRF SPI ports raise only
COMPLETED, so nothing loses a completion.

Peer data storage
-----------------

The error path in peer_data_delete_process decremented delete_counter,
sent PM_EVT_PEER_DELETE_FAILED and broke the inner loop only. Control
then fell into the success path below, which decremented a second time,
sent PM_EVT_PEER_DELETE_SUCCEEDED for the same peer and freed its id
while its records were still stored. Three faults from one misplaced
break: failure then success for one peer, an id handed back for reuse
with data under it, and a double decrement that drove delete_counter
negative. Since pds_peer_id_free only starts the loop on the 0 to 1 step,
that last one stopped every later delete request for the life of the run.
The decrement now happens once, after the inner loop. A failed peer keeps
its id and stays marked deleted, so the next delete request retries it.

find_next_data_entry_in_peer treated a BtPdsRead result of 0 as absent.
BtPdsWrite accepts a zero length value deliberately, so 0 means present
and empty while absent is -ENOENT. The sweep skipped such a record and
the peer id was freed with one of its entries still in the store.

SDC bond persistence
--------------------

BtSmpBondSave refused any save while one was staged, so a second save
during one pairing kept the older record and wrote that instead of the
current state of the bond. Refusing is only right when the staged record
belongs to a different slot, which is the case one buffer cannot hold;
that one is now counted. A full application event queue no longer
discards the record either: it stays staged and the next save queues the
handler again. BondSaveHandler takes the record and releases the staging
buffer before the write rather than after, since BtSmpBondSave runs at
the priority the stack dispatches at and could otherwise overwrite the
record being written.

UART
----

TxDropCnt was computed as Datalen - cnt in three ports. By that point the
loop has decremented Datalen and incremented cnt by the same amounts, so
the drop is Datalen alone, and subtracting cnt underflowed the unsigned
counter whenever more was accepted than remained. The rtry test also only
ran when the FIFO refused, so with a large FIFO the counter read 0
whatever was lost.

UARTvprintf discarded characters in two places and counted neither: a
line longer than SPRT_BUFFER_SIZE loses its tail in the format buffer,
and the five attempt loop abandons the rest of the line when UARTTx keeps
answering 0, which happens when DeviceIntrfStartTx cannot take the
interface and the driver is never entered. Both now go to TxDropCnt.

Found while tracking output that went missing on nRF54L15 at 1 Mbaud with
no flow control. Full accounting from printf to DMA.TX.AMOUNT showed
every byte present, and the same build is clean at 115200, so the loss is
on the link and not in the software. A working drop count would have said
so at the start.

Host test
---------

The transfer token was reverted from DevIntrf_t and Nvm::IntrfEvent, but
nvm_test.cpp still wrote pDev->XferToken and called IntrfEvent with two
arguments, so the documented host build did not compile. The NOR mock is
also fixed: NorStartRx keeps the frame on purpose, so a status read
reached TxSrData with

+++ src/bluetooth/bt_pds.cpp
@@ -712,7 +712,11 @@ static int ScanAll(void)
 		}
 	}
 
-	s_NextSeq = s_SeqFound ? (uint16_t)(s_MaxSeq + 1U) : 0U;
+	// Seq is 32 bit in the record header and both comparisons that pick the
+	// newest record use the full width. Narrowing it here restarted the
+	// counter low after 65535 records, and a record written afterwards then
+	// lost to an older one in FindLatest.
+	s_NextSeq = s_SeqFound ? s_MaxSeq + 1U : 0U;
 	return 0;
 }
